+++ pkg/secrets/repositoryserver/repository_password.go
@@ -18,16 +18,18 @@ import (
 	"fmt"
 
 	"github.com/pkg/errors"
-	v1 "k8s.io/api/core/v1"
+	corev1 "k8s.io/api/core/v1"
 )
 
 var ErrValidate = fmt.Errorf("validation Failed")
 
+var _ RepositoryServerSecret = &RepositoryPassword{}
+
 type RepositoryPassword struct {
-	password *v1.Secret
+	password *corev1.Secret
 }
 
-func NewRepoPassword(secret *v1.Secret) *RepositoryPassword {
+func NewRepoPassword(secret *corev1.Secret) *RepositoryPassword {
 	return &RepositoryPassword{
 		password: secret,
 	}
